refactor(user): 패키지 구조 변경 및 1차 리팩토링
- 멤버, 매개변수 final 작업
