+++ app/src/main/java/com/codesoom/assignment/user/application/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.user.exception;
+package com.codesoom.assignment.user.application.exception;
 
 /**
  * 회원을 찾지 못했을 때 던지는 예외입니다. <br>
