+++ app/src/main/java/com/codesoom/assignment/user/application/port/command/UserCreateRequest.java
@@ -1,6 +1,6 @@
-package com.codesoom.assignment.user.domain.port.command;
+package com.codesoom.assignment.user.application.port.command;
 
-import com.codesoom.assignment.user.repository.User;
+import com.codesoom.assignment.user.domain.User;
 
 public interface UserCreateRequest {
     String getEmail();
