+++ app/src/test/java/com/codesoom/assignment/domain/AuthenticationServiceTest.java
@@ -7,8 +7,8 @@ import com.codesoom.assignment.auth.application.port.in.AuthenticationUseCase;
 import com.codesoom.assignment.common.utils.JwtUtil;
 import com.codesoom.assignment.role.repository.Role;
 import com.codesoom.assignment.role.repository.RoleRepository;
-import com.codesoom.assignment.user.repository.User;
-import com.codesoom.assignment.user.repository.UserRepository;
+import com.codesoom.assignment.user.application.port.out.UserRepository;
+import com.codesoom.assignment.user.domain.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
